+++ Spring/Globa/src/main/java/org/y2k2/globa/application/study/dto/request/RequestStudyDto.java
@@ -6,5 +6,5 @@ import jakarta.validation.constraints.NotNull;
 public record RequestStudyDto(
         @NotNull(message = "공부 시간은 필수입니다.")
         @Min(0)
-        Long studyTime
+        Integer studyTime
 ) {}
