+++ Spring/Globa/src/main/java/org/y2k2/globa/infrastructure/persistence/study/projection/StudyTimeProjection.java
@@ -1,8 +1,6 @@
 package org.y2k2.globa.infrastructure.persistence.study.projection;
 
-import java.time.LocalDateTime;
-
 public interface StudyTimeProjection {
-    Long getTotalStudyTime();
+    Integer getTotalStudyTime();
     String getCreatedTime();
 }
